+++ app/controllers/services/opinions_controller.rb
@@ -16,6 +16,7 @@ class Services::OpinionsController < ApplicationController
 
   def index
     @service = Service.friendly.find(params[:service_id])
+    authorize(ServiceContext.new(@service, params.key?(:from) && params[:from] === "backoffice_service"), :show?)
     @related_services = @service.related_services
     @related_services_title = "Related resources"
     @service_opinions = ServiceOpinion.includes(project_item: :offer).