+++ app/controllers/services/questions_controller.rb
@@ -4,6 +4,7 @@ class Services::QuestionsController < ApplicationController
   def new
     @question = Service::Question.new
     @service = Service.friendly.find(params[:service_id])
+    authorize(ServiceContext.new(@service, params.key?(:from) && params[:from] === "backoffice_service"), :show?)
 
     respond_to do |format|
       format.js { render_modal_form }
@@ -12,14 +13,15 @@ class Services::QuestionsController < ApplicationController
 
   def create
     @service = Service.friendly.find(params[:service_id])
+    authorize(ServiceContext.new(@service, params.key?(:from) && params[:from] === "backoffice_service"), :show?)
     user = current_user
     @question = Service::Question.new(author: user&.full_name || params[:service_question][:author],
-                                     email: user&.email || params[:service_question][:email],
-                                     text: params[:service_question][:text],
-                                     service: @service)
+                                      email: user&.email || params[:service_question][:email],
+                                      text: params[:service_question][:text],
+                                      service: @service)
     respond_to do |format|
       if @question.valid? && verify_recaptcha(model: @question, attribute: :verified_recaptcha)
-        @service.public_contacts.each  do |contact|
+        @service.public_contacts.each do |contact|
           ServiceMailer.new_question(contact.email, @question.author, @question.email,
                                      @question.text, @service).deliver_later
         end
@@ -36,10 +38,10 @@ class Services::QuestionsController < ApplicationController
       render "layouts/show_modal",
              content_type: "text/javascript",
              locals: {
-                title: "Ask provider",
-                action_btn: t("simple_form.labels.question.new"),
-                form: "services/questions/form",
-                form_locals: { service: @service, question: @question }
+               title: "Ask provider",
+               action_btn: t("simple_form.labels.question.new"),
+               form: "services/questions/form",
+               form_locals: { service: @service, question: @question }
              }
     end
 end