+++ app/policies/ordering_configuration_policy.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class OrderingConfigurationPolicy < ApplicationPolicy
-  def show?
-    record.administered_by?(user)
-  end
-end